+++ changelog.md
@@ -1,3 +1,4 @@
 - Fixed error when launching the game
 - Fixed zoomed images not being zoomed in
 - Fixed images rendered darker
+- Fixed image shading